app/requirements.txt
streamlit
psutil
watchdog
requests
pandas
python-dotenv